requirements.txt
# PDF & DOCUMENT PROCESSING
PyMuPDF==1.23.8
pdf2image==1.16.3
pdfplumber==0.10.3
camelot-py[cv]==0.11.0

# OCR
pytesseract==0.3.10

# IMAGE PROCESSING
Pillow==10.1.0
opencv-python==4.8.1.78

# DATA HANDLING
numpy==1.26.2
pandas==2.1.4

# MACHINE LEARNING
scikit-learn==1.3.2

# DEEP LEARNING (CPU ONLY – INFERENCE)
torch==2.1.0+cpu
torchvision==0.16.0+cpu
timm==0.9.12

# UTILITIES
tqdm==4.66.1
python-dotenv==1.0.0


# pip install -r requirements.txt --index-url https://download.pytorch.org/whl/cpu
